refactor: use store instead of context for modal management (#133)
- removes the modal context
- aligns modal state with zustand usage
- adds the nice fade in for modals that was dropped
