+++ src/index.css
@@ -29,18 +29,6 @@
       opacity: 1;
     }
   }
-
-  @keyframes modalFadeIn {
-    from {
-      opacity: 0;
-      transform: translateY(-20px);
-    }
-
-    to {
-      opacity: 1;
-      transform: translateY(0);
-    }
-  }
 }
 
 input {
@@ -196,4 +184,20 @@ body {
 
 .message.outgoing .file-message {
   background-color: rgba(255, 255, 255, 0.1);
+}
+
+.modalFadeIn {
+  animation: modalFadeIn 0.2s ease;
+}
+
+@keyframes modalFadeIn {
+  from {
+    opacity: 0;
+    transform: translateY(-20px);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
